Fix file tests for shortcuts
